+++ src/main/java/com/example/bookstoreapp/repository/DoubleSpecificationProvider.java
@@ -1,10 +1,11 @@
 package com.example.bookstoreapp.repository;
 
+import java.math.BigDecimal;
 import org.springframework.data.jpa.domain.Specification;
 
 public interface DoubleSpecificationProvider<T> {
 
     String getKey();
 
-    Specification<T> getSpecification(Double[] params);
+    Specification<T> getSpecification(BigDecimal[] params);
 }
